refactor: LottoGenerator 생성
LottoGenerator는 로또 생성과 관련된 기능을 모두 포함합니다.
